Updating the project statement
As advised by TA Ms Woo: 'I believe the program features listed in the readme file are feasible and organised, but one point - it might be good to indicate which of them are required functions and which ones are new functions your group have brainstormed. This can help make things clearer. The problem setting is well-defined, but regarding problem statement I think you may try to give more context. Why do we need to "optimize the management process of commodities in warehouses and shops"? You may try adding a few more lines about the status quo of management systems.'